chore: refactor directory structure
