+++ packages/merchant-server/index.ts
@@ -4,14 +4,8 @@ import dotenv from 'dotenv';
 
 dotenv.config();
 
-import app, { logger } from './app';
+import app from './app';
 
 const port = Number(process.env.PORT) || 8000;
 
-app.listen({ port }, (err) => {
-  if (err) {
-    logger.error(err);
-  } else {
-    logger.info(`merchant-server start on port ${port}`);
-  }
-});
+app.listen({ port });
