+++ packages/merchant-server/knexfile.ts
@@ -8,7 +8,10 @@ dotenv.config({ path: '../merchant-server/.env' });
 export const development = {
   client: 'pg',
   connection: process.env.DB_CONNECTION_STRING,
+  migrations: {
+    directory: './infra/db-migrations',
+  },
   seeds: {
-    directory: './migrations/seeds',
+    directory: './infra/db-migrations/seeds',
   },
 };
